refactor: move intrinsic types to simple
